// src/resources/chat/chat.types.ts
export type UniqueId = string;
export type messageId = UniqueId;

// ========== 变量 ==========

/**
 * 变量变更应当是一段可执行的代码片段。
 * 代码应包含一个 `async main(oldValue)` 函数。
 */
export interface VariableChange {
  /**
   * 包含 `async main(oldValue)` 函数定义的代码字符串。
   * @example
   * "async main(data) {
   *    data.gold += 100;
   *    return data;
   *  }"
   */
  code: string;
}

// ========== 区间 ==========
export type IntervalEndCondition =
  | { type: "length"; value: number }
  | { type: "anchor"; anchor: messageId };

export type ActivationCondition =
  | { type: "relative_index"; value: number }
  | { type: "next_message_with_role"; role: "user" | "assistant" };

export interface AttachedIntervalDef {
  id: UniqueId;
  source: string;
  type: string; // A type of "hidden" will be used for the new mechanism
  content: Record<string, any>;
  endCondition: IntervalEndCondition;
  removeOnAnchorDeletion?: boolean;
  onUnmatchedEnd?: "leak" | "truncate_at_leaf" | "discard";
  activationCondition?: ActivationCondition;
}

export interface AttachedEndMarkerDef {
  endsIntervalById?: UniqueId[];
  endsIntervalByType?: string[];
}

// ========== 消息相关 ==========

export interface ToolCallResult {
  type: "tool-result";
  toolCallId: string;
  toolName: string;
  input: any;
  output: any;
}

// 为 steps 数组定义本地步骤类型
export interface LocalStep {
  name: string;
  message: string;
}

export interface MetaGenerateInfo {
  modelName: string;
  timeInfo: {
    start: string;
    timeUsed?: number;
  };
  renderInfo: {
    // preset自身name属性
    usedPresetName: string;
    /**
     * 角色的资源路径，用于获取头像
     * @example
     * import {useFileSystemStore} from "@/features/fileSystem";
     * const fs = useFileSystemStore()
     * fs.getAvatarUrl(message.content.metaGenerateInfo.renderInfo.characterFilePath)
     */
    characterFilePath: string;
    // ctx.CHARACTER.name
    characterName: string;
    // preset自身needToBakeregex属性
    bakedRegexReplace: {
      find_regex: string;
      replace_string: string;
      applyOn: "rendering" | "generating";
    }[];
  };
  // 存储该消息的向量化结果
  // 键是 modelId (例如 "openai/text-embedding-3-small")
  // 值是向量数组 (number[])
  embedding?: Record<string, number[]>;
  // 置空，之后多步骤生成使用
  steps: (LocalStep | ToolCallResult)[];

  // 置空，供之后使用
  /** 此版本触发的变量变更 */
  variableChanges?: VariableChange[];
  // 从中间值中提取并附加
  /** 锚定在此版本上的区间定义 */
  attachedIntervals?: AttachedIntervalDef[];
  // 置空，供之后使用
  /** 锚定在此版本上的结束标记 */
  attachedEndMarker?: AttachedEndMarkerDef;
  // 置空，供之后使用
  /** 附加内容部分 */
  additionalParts?: AdditionalParts[];
}
// ...
export type AdditionalParts = TextPart | ImagePart | FilePart;
// ...
export interface ApiReadyMessage {
  id?: UniqueId;
  role: role;
  content: string;
  metaGenerateInfo?: MetaGenerateInfo;
}